Add tests for eval and resume too
