+++ emloop/cli/__init__.py
@@ -1,9 +1,9 @@
 from .ls import list_train_dirs
 
 from .train_fn import train
-from .resume import resume
+from .resume_fn import resume
 from .grid_search import grid_search, _build_grid_search_commands
-from .eval import evaluate
+from .eval_fn import evaluate
 from .dataset import invoke_dataset_method
 from .util import find_config, validate_config, fallback
 from .args import get_emloop_arg_parser
