+++ emloop/cli/resume_fn.py
@@ -34,8 +34,7 @@ def resume(config_path: str, restore_from: Optional[str], cl_arguments: Iterable
 
         logging.debug('\tLoaded config: %s', config)
 
-        emloop_training = create_emloop_training(
-            config=config, output_root=output_root, restore_from=restore_from)
+        emloop_training = create_emloop_training(config, output_root, restore_from)
         emloop_training.main_loop.run_training()
     except (Exception, AssertionError) as ex:  # pylint: disable=broad-except
         logging.error('Resume failed')
